+++ orc/orcemulateopcodes.c
@@ -4896,7 +4896,7 @@ emulate_convlf (OrcOpcodeExecutor *ex, int offset, int n)
     /* 0: loadl */
     var32 = ptr4[i];
     /* 1: convlf */
-    var33.f = var32.i;
+    var33.f = (float) var32.i;
     /* 2: storel */
     ptr0[i] = var33;
   }
@@ -5279,7 +5279,7 @@ emulate_convdl (OrcOpcodeExecutor *ex, int offset, int n)
     /* 1: convdl */
     {
        int tmp;
-       tmp = var32.f;
+       tmp = (int) var32.f;
        if (tmp == 0x80000000 && !(var32.i & ORC_UINT64_C(0x8000000000000000))) tmp = 0x7fffffff;
        var33.i = tmp;
     }
